Add module for working with Yaml
Working with NimYAML is unpleasant due to the number of tags that it
spits out in output and due to the APIs not matching how I wanted to use
them.

This module currently still uses NimYAML for parsing, but provides
a YNode type to work with Yaml objects at runtime and an alternate
toString() serialization proc.

Future expansions:
  code generation for ofYaml/toYaml based on type information
    - should support polymorphism and variants